// src/interfaces/DB/Card.ts
export default interface Card {
  name: string
  code: string
  rarity: string
  expansion: string
  side: string
  type: string
  color: string
  level: string
  cost: string
  power: string
  soul: string
  trigger: string[]
  attributes: string[]
  ability: string[]
  flavor_text: string
  set: string
  release: string
  sid: string
  image: string
}